requirements.txt
PyYAML
ConfigParser
pathlib
python-dotenv
json
logging
unittest
mock
coverage